Different auth for Reconnect & agents + check if agent is the right one before accessing content
